R01/requirements.txt
MySQL-python==1.2.5
numpy==1.15.4
pandas==0.23.4
python-dateutil==2.7.5
pytz==2018.7
scikit-learn==0.20.2
scipy==1.1.0
six==1.12.0
sklearn==0.0
